Fix significant bug in serial talker.
When writeable events are enabled all the time, it just spams the
event handler, which uses 100% CPU. The unit tests never caught
this because they happen too fast for the spike to be noticeable.

Anyway, I fixed it by only enabling the writeable event when
after a write doesn't actually complete fully.